Created new callstack_instrument test
