+++ test/test_http_instrumentation.py
@@ -12,7 +12,6 @@ import pytest
 
 from ..automation import CommandSequence, TaskManager
 from ..automation.utilities import db_utils
-from ..automation.utilities.platform_utils import parse_http_stack_trace_str
 from . import utilities
 from .openwpmtest import OpenWPMTest
 
@@ -451,46 +450,6 @@ HTTP_SERVICE_WORKER_REQUESTS = {
     ),
 }
 
-# HTTP request call stack instrumentation
-# Expected stack frames
-HTTP_STACKTRACE_TEST_URL = utilities.BASE_TEST_URL + "/http_stacktrace.html"
-STACK_TRACE_INJECT_IMAGE =\
-    "inject_image@" + HTTP_STACKTRACE_TEST_URL + ":18:7;null\n"\
-    "inject_all@" + HTTP_STACKTRACE_TEST_URL + ":22:7;null\n"\
-    "onload@" + HTTP_STACKTRACE_TEST_URL + ":1:1;null"
-
-RAWGIT_HTTP_STACKTRACE_TEST_URL = "https://gist.githubusercontent.com/gunesacar/b927d3fe69f3e7bf456da5192f74beea/raw/8d3e490b5988c633101ec45ef1443e61b1fd495e/inject_pixel.js"  # noqa
-# https://gist.github.com/gunesacar/b927d3fe69f3e7bf456da5192f74beea
-STACK_TRACE_INJECT_PIXEL =\
-    "inject_pixel@" + RAWGIT_HTTP_STACKTRACE_TEST_URL + ":4:3;null\n"\
-    "null@" + RAWGIT_HTTP_STACKTRACE_TEST_URL + ":6:1;null"
-
-STACK_TRACE_INJECT_JS =\
-    "inject_js@" + HTTP_STACKTRACE_TEST_URL + ":13:7;null\n"\
-    "inject_all@" + HTTP_STACKTRACE_TEST_URL + ":21:7;null\n"\
-    "onload@" + HTTP_STACKTRACE_TEST_URL + ":1:1;null"
-
-HTTP_STACKTRACES = set((STACK_TRACE_INJECT_IMAGE,
-                        STACK_TRACE_INJECT_PIXEL,
-                        STACK_TRACE_INJECT_JS))
-# parsed HTTP call stack dict
-CALL_STACK_INJECT_IMAGE =\
-    [{"func_name": "inject_image",
-      "filename": HTTP_STACKTRACE_TEST_URL,
-      "line_no": "18",
-      "col_no": "7",
-      "async_cause": "null"},
-     {"func_name": "inject_all",
-      "filename": HTTP_STACKTRACE_TEST_URL,
-      "line_no": "22",
-      "col_no": "7",
-      "async_cause": "null"},
-     {"func_name": "onload",
-      "filename": HTTP_STACKTRACE_TEST_URL,
-      "line_no": "1",
-      "col_no": "1",
-      "async_cause": "null"}]
-
 BASE_PATH = os.path.dirname(os.path.realpath(__file__))
 
 
@@ -622,37 +581,6 @@ class TestHTTPInstrument(OpenWPMTest):
             dst = row['new_request_url'].split('?')[0]
             observed_records.add((src, dst))
         assert HTTP_CACHED_REDIRECTS == observed_records
-
-    def test_http_stacktrace(self):
-        test_url = utilities.BASE_TEST_URL + '/http_stacktrace.html'
-        db = self.visit(test_url, sleep_after=3)
-        rows = db_utils.query_db(db, (
-            "SELECT url, req_call_stack FROM callstacks"))
-        observed_records = set()
-        for row in rows:
-            print(row)
-            url, stacktrace = row
-            if (url.endswith("inject_pixel.js")
-                    or url.endswith("test_image.png")  # noqa: W503
-                    or url.endswith("Blank.gif")):  # noqa: W503
-                observed_records.add(stacktrace)
-        assert HTTP_STACKTRACES == observed_records
-
-    def test_parse_http_stack_trace_str(self):
-        stacktrace = STACK_TRACE_INJECT_IMAGE
-        stack_frames = parse_http_stack_trace_str(stacktrace)
-        assert stack_frames == CALL_STACK_INJECT_IMAGE
-
-    # TODO: webext instrumentation doesn't support req_call_stack yet.
-    # def test_http_stacktrace_nonjs_loads(self):
-    #     # stacktrace should be empty for requests NOT triggered by scripts
-    #     test_url = utilities.BASE_TEST_URL + '/http_test_page.html'
-    #     db = self.visit(test_url, sleep_after=3)
-    #     rows = db_utils.query_db(db, (
-    #         "SELECT url, req_call_stack FROM http_requests"))
-    #     for row in rows:
-    #         _, stacktrace = row
-    #         assert stacktrace == ""
 
     def test_javascript_saving(self, tmpdir):
         """ check that javascript content is saved and hashed correctly """
